+++ NEWS.md
@@ -1,6 +1,7 @@
 # FSA 0.8.8 ongoing
 * `growthFunShow()`: Modified.  Added Pauly et al. (1992) seasonal cessation function.
 * `vbFuns()`: Modified.  Added Pauly et al. (1992) seasonal cessation function.
+* `vbStarts()`: Modified.  Added `fixed=` so that the user can define some of the starting values.  Added Pauly et al. (1992) seasonal cessation function.  Added tests for `fixed=`.
 
 # FSA 0.8.7 7-May-16
 * Compiled under R v3.3.0.